feat:优化 plugin 代码
