+++ core/src/plugins/index.ts
@@ -1,4 +1,4 @@
-export { default as SSRWebpackRunPlugin } from './webpackRun';
+export { default as SSRWebpackRunPlugin } from './webpackRun/index';
 export { default as SSRWebpackPlugin } from './plugin';
 export * from "./utils/module"
 export * from "./utils/plugins"
