+++ example/basic-plugins/.kktrc.js
@@ -2,7 +2,7 @@ import path from 'path';
 import { SSRWebpackRunPlugin, SSRWebpackPlugin, getRemoveHtmlTemp } from '@kkt/ssr/lib/plugins';
 process.env.FAST_REFRESH = 'false';
 process.env.GENERATE_SOURCEMAP = 'false';
-export default (conf) => {
+export default (conf, evn) => {
   conf.plugins.push(new SSRWebpackRunPlugin());
   conf.plugins = getRemoveHtmlTemp(conf.plugins)
   conf.stats = {
